+++ src/main/java/com/dtone/dvs/util/HttpConstants.java
@@ -1,0 +1,10 @@
+package com.dtone.dvs.util;
+
+public class HttpConstants {
+
+  public static final String UNDEFINED_HEADER = "-1";
+
+  public static final String USER_AGENT = "User-Agent";
+  public static final String CONTENT_TYPE = "Content-Type";
+  public static final String APPLICATION_JSON = "application/json";
+}