+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## [2.0.0] - 2022-09-04
+### Added
+- Migration to Java 17
+- Migration to Java 11 HttpClient
+- Migration to asynchronous calls
+- Upgrade dependencies
+
 ## [1.0.0] - 2019-12-08
 ### Added
 - API Discovery, Transactions, Account and Lookup interfaces